+++ pkgs/cupertino_http/lib/src/cupertino_client.dart
@@ -15,6 +15,28 @@ import 'cupertino_api.dart';
 
 final _digitRegex = RegExp(r'^\d+$');
 
+const _nsurlErrorCancelled = -999;
+
+/// A [ClientException] generated from an [NSError].
+class NSErrorClientException extends ClientException {
+  final NSError error;
+
+  NSErrorClientException(this.error, [Uri? uri])
+      : super(error.localizedDescription.toDartString(), uri);
+
+  @override
+  String toString() {
+    final b = StringBuffer(
+        'NSErrorClientException: ${error.localizedDescription.toDartString()} '
+        '[domain=${error.domain.toDartString()}, code=${error.code}]');
+
+    if (uri != null) {
+      b.write(', uri=$uri');
+    }
+    return b.toString();
+  }
+}
+
 /// This class can be removed when `package:http` v2 is released.
 class _StreamedResponseWithUrl extends StreamedResponse
     implements BaseResponseWithUrl {
@@ -33,12 +55,15 @@ class _StreamedResponseWithUrl extends StreamedResponse
 class _TaskTracker {
   final responseCompleter = Completer<URLResponse>();
   final BaseRequest request;
-  final responseController = StreamController<Uint8List>();
+  final StreamController<Uint8List> responseController;
+
+  /// Whether the response stream subscription has been cancelled.
+  bool responseListenerCancelled = false;
   final HttpClientRequestProfile? profile;
   int numRedirects = 0;
   Uri? lastUrl; // The last URL redirected to.
 
-  _TaskTracker(this.request, this.profile);
+  _TaskTracker(this.request, this.responseController, this.profile);
 
   void close() {
     responseController.close();
@@ -167,9 +192,14 @@ class CupertinoClient extends BaseClient {
   static void _onComplete(
       URLSession session, URLSessionTask task, NSError? error) {
     final taskTracker = _tracker(task);
-    if (error != null) {
-      final exception = ClientException(
-          error.localizedDescription.toDartString(), taskTracker.request.url);
+    // The task will only be cancelled if the user calls
+    // `StreamedResponse.stream.cancel()`, which can only happen if the response
+    // has already been received. Therefore, it is safe to handle task
+    // cancellation errors as if the response completed normally.
+    if (error != null &&
+        !(error.domain.toDartString() == 'NSURLErrorDomain' &&
+            error.code == _nsurlErrorCancelled)) {
+      final exception = NSErrorClientException(error, taskTracker.request.url);
       if (taskTracker.profile != null &&
           taskTracker.profile!.requestData.endTime == null) {
         // Error occurred during the request.
@@ -184,6 +214,7 @@ class CupertinoClient extends BaseClient {
         taskTracker.responseCompleter.completeError(exception);
       }
     } else {
+      assert(error == null || taskTracker.responseListenerCancelled);
       assert(taskTracker.profile == null ||
           taskTracker.profile!.requestData.endTime != null);
 
@@ -199,6 +230,7 @@ class CupertinoClient extends BaseClient {
 
   static void _onData(URLSession session, URLSessionTask task, NSData data) {
     final taskTracker = _tracker(task);
+    if (taskTracker.responseListenerCancelled) return;
     taskTracker.responseController.add(data.toList());
     taskTracker.profile?.responseData.bodySink.add(data.toList());
   }
@@ -338,7 +370,13 @@ class CupertinoClient extends BaseClient {
     // This will preserve Apple default headers - is that what we want?
     request.headers.forEach(urlRequest.setValueForHttpHeaderField);
     final task = urlSession.dataTaskWithRequest(urlRequest);
-    final taskTracker = _TaskTracker(request, profile);
+    final subscription = StreamController<Uint8List>(onCancel: () {
+      final taskTracker = _tasks[task];
+      if (taskTracker == null) return;
+      taskTracker.responseListenerCancelled = true;
+      task.cancel();
+    });
+    final taskTracker = _TaskTracker(request, subscription, profile);
     _tasks[task] = taskTracker;
     task.resume();