+++ pkgs/cupertino_http/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.2.0
+
+* Cancel requests when the response stream is cancelled.
+* Add a new exception type `NSErrorClientException` that contains the
+  `NSError` associated with the failure.
+
+## 2.1.1
+
+* Support `package:web_socket` 1.0.0.
+
 ## 2.1.0
 
 * Remove some unnecessary native code.